fix(sso): respect X-Forwarded-* headers for redirect URL. Fixes #13031
When running Argo in plain text mode behind a TLS termination proxy with
SSO enabled, you will always be redirected to the workflows page after
logging in, regardless of which page you came from. This is happening
because `server/auth/sso/sso.go` is using the protocol and host of the
request to determine the redirect URL, which can be different than the
original request when a proxy is used.

`oauth2-proxy` had a similar issue, which they fixed in
https://github.com/oauth2-proxy/oauth2-proxy/pull/957 by reading from
the
[X-Forwarded-Host](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-Host)
and
[X-Forwarded-Proto](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-Proto)
headers, if set. This applies the same fix. The only major difference is
that `oauth2-proxy` only reads those headers when the `--reverse-proxy`
option is set, but I didn't add a corresponding option here. The only
reason I can think of for doing so is to preserve
backwards-compatibility, since it's theoretically possible this could
break someone's setup. I can't think of any real-world situations where
that can happen, but let me know if I should put this behind a option
and I'm happy to do so.

Tested using `make start UI_SECURE=true PROFILE=sso` and changing this
line: https://github.com/argoproj/argo-workflows/blob/29ea3518a30b2b605290d0ba38f95bfb75a2901f/manifests/quick-start/sso/overlays/workflow-controller-configmap.yaml#L15
to `redirectUrl: https://localhost:8080/oauth2/callback`. I tried to see
if I could automate this, but I couldn't figure out a clean way of doing
so.

Signed-off-by: Mason Malone <mmalone@adobe.com>