fix(server): respect workflowDefaults when validating workflows. Fixes #10946
Submitting a workflow with a variable reference that points to a label
or annotation that comes from `workflowDefaults` will always fail with
an error like the following:
```
msg="Failed to submit workflow: rpc error: code = InvalidArgument desc = templates.whalesay-template: failed to resolve {{workflow.labels.arg-name}}
```

This is happening because the workflow is merged with `workflowDefaults`
at runtime by
https://github.com/argoproj/argo-workflows/blob/2dac1266b446d059ac6df78afe3d09a6ddb3af4b/workflow/controller/controller.go#L1375

which happens before the workflow is validated by `ValidateWorkflow()`.
Therefore, we need to explicitly pass in the `workflowDefaults` (if any)
to `ValidateWorkflow()` so it can merge in any labels/annotations before
validating templates.

Signed-off-by: Mason Malone <mmalone@adobe.com>