refactor(cli): consolidate and fix `--output` validation
This is a follow-up to
https://github.com/argoproj/argo-workflows/pull/13656#pullrequestreview-2341800427.
Many of the CLI commands take an `--output` flag, but this wasn't
being validated, and there was a lot of duplication. The duplication has
led to docs getting out-of-date, e.g. `argo cluster-template get --help`
says `Output format. Must be one of: json|yaml|wide`, but it accepts
`name` too.

This introduces a simple `EnumFlagValue` type to represent enum flags
like `--output` and changes all commands with an `--output` flag to use
it. This will cause Cobra to give a helpful error if you pass an invalid
value, e.g.:
```
$ ./dist/argo list -o INVALID
Error: invalid argument "INVALID" for "-o, --output" flag: One of: name|json|yaml|wide
```

Signed-off-by: Mason Malone <mmalone@adobe.com>