fix!: consolidate cronworkflow variables
Issues #13474 and #12305 both introduced expressions into
CronWorkflows as new features for 3.6.

Motivation

The expressions for `when` include the prefix `cronworkflow.`. Those
for `stopStrategy.condition` do not.

Unify both of these under `cronworkflow.` and allow both sets of
variables to be used in both expressions. This simplifies the
documentation and is less surprising.

stopExpression's `condition` is mostly not a field name used, so in
agreement with agilgur5 I've renamed it to `expression`.

Modifications

Renamed `stopStrategy.condition` to `stopStrategy.expression`

Created a single function to generate all the variables for both
expressions and use that instead of the two disparate mechanisms.

Modified some code comments to be godoc compliant.

Verification

Existing tests still pass with appropriate amendments to the test cases

Note

This is a breaking change vs 3.6-rc1 and rc2 for users of stopStrategy
only.

Signed-off-by: Alan Clucas <alan@clucas.org>