test: increase timeout for e2e-tests
Occasionally, tests are being canceled due to exceeding the 30 minute
timeout. Example: https://github.com/argoproj/argo-workflows/actions/runs/11358993638/job/31594490093

While there are optimizations we can make to the test suite to mitigate
this, sometimes resource contention causes things to slow to a crawl for
reasons outside our control. In the example above, it spent 9m15s on
"Free up unused disk space" step, which normally takes ~12s. The only
thing that step is doing is running `rm -rf`.

This removes the timeout so it uses the default of 360 minutes (source:
https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#jobsjob_idtimeout-minutes).
This is what ArgoCD does (source:
https://github.com/argoproj/argo-cd/blob/049ba0aab3965f3fa73447ae2db08f3a2afd95a0/.github/workflows/ci-build.yaml#L398), and E2E tests for ArgoCD PRs often take ~1hour (example: https://github.com/argoproj/argo-cd/actions/runs/11352918460/job/31576914061).

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>