ci: login to Docker Hub to avoid rate limiting
Partial fix for #10807.

Builds are occasionally failing while pulling images from Docker Hub due
to rate limiting, e.g. https://github.com/argoproj/argo-workflows/actions/runs/11564257560/job/32189242898:
>  Oct 28 23:30:52 fv-az802-461 k3s[2185]: E1028 23:30:52.151698    2185 kuberuntime_image.go:53] "Failed to pull image" err="Error response from daemon: toomanyrequests: You have reached your pull rate limit. You may increase the limit by authenticating and upgrading: https://www.docker.com/increase-rate-limit" image="minio/minio:RELEASE.2022-11-17T23-20-09Z"

As explained in the error, logging in increases the limit. We're already
doing this in the "Release" workflow, so this copies that over.

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>