fix: don't print help for non-validation errors. Fixes #13826
https://github.com/argoproj/argo-workflows/pull/13656 changed all CLI
commands to use `RunE` instead of `Run` and [cobra
validators](https://cobra.dev/#positional-and-custom-arguments). The
default behavior with Cobra is to print the help text for all errors,
which can be tedious to scroll through.

This changes the CLI to only print the help text for argument validation
errors, which should match the previous behavior.

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>