fix: consistent variable substitution for configMapKeyRef. Fixes #13890
Support for subsituting variables in `configMapKeyRef` was added in
https://github.com/argoproj/argo-workflows/pull/7542 by manually parsing
templates out of the string instead of using the standard
`template.Replace()` functionality. Since the templates have the same
syntax as simple template tags (e.g. `{{workflow.parameters.cm-name}}`),
users reasonably expect it to work the same, but it doesn't, which leads
to confusion.

This replaces the manual parsing code with the standard
`template.Replace()` function that everything else uses.

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>