fix(ui): handle parsing errors properly
https://github.com/argoproj/argo-workflows/pull/13915 introduced a
regression where edits made in an object editor that introduce syntax
errors cause a full screen error. The editor uses the following
`try-catch` block to handle parse errors, but the `catch` block no
longer works because the parsing logic was moved to `reducer()`, which
is executed async:
https://github.com/argoproj/argo-workflows/blob/main/ui/src/shared/components/object-editor.tsx#L117-L122

This moves the parsing logic to the `setObject()` function, which means
errors are now propagated to the caller.

Tested by going http://localhost:8080/workflow-templates, creating a new
`WorkflowTemplate`, and making edits.

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>