ci: run API tests on PostgreSQL and add test for unicode escaping
The workflow archive logic at `persist/sqldb/workflow_archive.go`
currently has no unit tests, so the only automated tests for it are the
E2E tests. But the E2E tests are only being run with MySQL, and there's
quite a bit of PostgreSQL-specific logic in there that's not being
tested.

This updates the CI to run `make test-api` with PostgreSQL in addition
to MySQL, and adds a test to verify Unicode escapes aren't being mangled
when archiving a workflow, which is intended to help with fixing
https://github.com/argoproj/argo-workflows/issues/13711

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>