ci: optimize tests by only building binaries once
All the E2E tests job run `make controller` to build the workflow
controller, and several also run `make cli` to build the CLI, which
takes ~2 minutes. It'll be faster to have a central "build binaries" job
that does that and uploads them as an artifact, which each E2E job can
then download.

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>