test: fix and optimize flaky cron tests. Partial fix for #10807
The cron tests have been some of the slowest and flakiest for awhile
now, mostly due to using `time.Sleep()` to wait a fixed amount of time.
The `TestCronSuite` function tries to workaround the flakiness by
running the suite at the 30 second mark, but this is only effective for the
first time in the suite.

This changes nearly all the cron tests to remove the `time.Sleep()`
calls and instead use the API to poll for changes. The one test I
couldn't fix to stop using `time.Sleep()` is `TestBasicForbid`, since I
don't know any way of verifying the `concurrencyPolicy` is respected.

Also, most of the cron fixture data wasn't specifying workflow metadata
properly using `workflowMetadata`, which meant they weren't being
cleaned up automatically, so I fixed that too.

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>