fix(api/jsonschema): use unchanging JSON Schema version
The current value of `$schema` in `api/jsonschema/schema.json` refers to
a schema version that changes over time, which means that there's a
possibility that the keywords used in the schema might change meaning
unexpectedly when a new version is released.

This PR fixes the schema version to the current latest version, avoiding
this potential issue by locking in the expected JSON Schema semantics
and meaning that any would-be interpreter does not have to make a
network request to find out what version is intended.

I checked that the new schema validates OK in a JSON schema checker.

A passing remark with regard to this comment in
`hack/api/jsonschema/main.go`:

```
"$id":     "http://workflows.argoproj.io/workflows.json", // don't really know what this should be
```

It seems relatively common to use a working URL for JSON schemas URIs so
this could potentially be changed to:

```
"$id":     "https://raw.githubusercontent.com/argoproj/argo-workflows/HEAD/api/jsonschema/schema.json",
```

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>