fix: gracefully handle invalid CronWorkflows and simplify logic. Fixes #14047
If you have a `CronWorkflow` that's missing both `schedules` and
`schedule`, or has `schedules` set to an empty array, you'll get an
error in the UI when visitng `/cron-workflows`. This appears to have
been introduced https://github.com/argoproj/argo-workflows/pull/12616.
You could argue this isn't a bug, since such `CronWorkflows` aren't
valid, but multiple users have reported this, so we should gracefully
handle it.

This updates the UI code that fetches `CronWorkflows` from the API to
normalize the response by setting `schedules` to an empty array if it
isn't present and automatically converting `schedule` to `schedules`.
Then, I deleted all the logic for handling `schedule`, since it's no
longer needed. Once we drop support for `schedule` completely, we could
delete the logic for converting that to `schedules`, but it's only a few
lines of code.

I also added some basic tests.

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>