fix: use quay.io for third-party images to avoid rate limiting
A major source of flaky CI builds for Argo Workflows is Docker Hub image limits. Here's an example:
> [port-forward] (starting) [pods/minio-769877b9db-6j8gt] (Waiting) ErrImagePull: container "main" is waiting: Error response from daemon: toomanyrequests: You have reached your unauthenticated pull rate limit. https://www.docker.com/increase-rate-limit

Source: https://github.com/argoproj/argo-workflows/actions/runs/13519739393/job/37776381312?pr=14215

On April 1st, Docker Hub is further reducing the limits, so I expect
this is going to get worse. This switches the images used in the
manifests (which are in turn used by the E2E tests) to use equivalent
images on quay.io:
* https://hub.docker.com/r/minio/minio -> https://quay.io/minio/minio
* https://hub.docker.com/r/kong/httpbin - quay.io/holos/mccutchen/go-httpbin
* https://hub.docker.com/_/mysql/tags?name=8 - https://quay.io/fedora/mysql-80
* https://hub.docker.com/_/postgres/tags?name=12-alpine - https://quay.io/repository/fedora/postgresql-12
* https://hub.docker.com/r/prom/prometheus/ - https://quay.io/repository/prometheus/prometheus

I also switched all images to use the manifest digest so we don't need
to worry about an image suddenly changing and causing all builds to fail
(and also partly for security).

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>