fix: silence "Ignoring duplicate key error". Fixes #14344
When offloading a workflow, it's possible for multiple workers to concurrently
call `Save()` with the same workflow, which leads to `ERROR: duplicate key value
violates unique constraint "argo_workflows_pkey"`` messages in the logs, along
with a stack trace. These messages are cluttering the logs and confusing users,
but duplicate key errors are harmless. When it detects a duplicate key error,
`Save()` will return the `version` hash, and `version` is part of the primary
key of the `argo_workflows` table, which means it's guaranteed to be identical
to the previously-inserted row.

This decreases the log level of that message to `DEBUG` so that it doesn't
clutter the logs. I thought about removing it entirely, but I figured it's worth
keeping just in case.

I wasn't able to reproduce the error locally, but I verified workflow offloading
works locally using `make PROFILE=postgres UI=true ALWAYS_OFFLOAD_NODE_STATUS=true`

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>