feat(workflowtemplate): evaluate expressions in workflow template parameters
This change introduces the capability to evaluate expressions within the default values of both global and template-level parameters in a . `WorkflowTemplate`. The result is inserted in the "value" field of the template.
It allows the UI user to get ready to use submit parameters, like dates, complex arithmetics and more.

Previously, default parameter values were treated as static strings. With this enhancement, string values that are encapsulated in `{{= ... }}` are now interpreted and evaluated as expressions.
The evaluation process is performed in two stages:
1. Global parameters are evaluated first. These parameters cannot reference other parameters.
2. Template-level parameters are then evaluated, with access to the values of the evaluated global parameters through the `workflow.parameters` context.

This allows for more powerful and dynamic , where template-level configurations can be derived from global settings. `WorkflowTemplates`

Signed-off-by: edaniel <edaniel@groupcls.com>