chore: bulk flip of example testing label #15002
This commit removes the label `workflows.argoproj.io/test` used to mark
examples as testable because now we want as contributors to be more
explicit marking  examples that DO NOT have tests, this
should help maintainers to ask for those tests as part of code review.

This commit adds `workflows.argoproj.io/no-test-broken` for genuinely
broken examples which can be fixed or that do not have test yet.

`workflows.argoproj.io/no-test-environment` for examples which need to
run in a different environment, e.g. need access to artifactory but this
commit does not mark those.

Related #15002

Signed-off-by: Gianluca Arbezzano <ciao@gianarb.it>