chore: make examples that pass tests to run #15002
Many examples could pass tests but they were marked as no-test for
various reasons.

I analyzed all the tests and I marked the one that did not pass as
`no-test-broken` or `-no-test-environment` according to why they were
not passing.

The most common issues for failing example were the image used as part
of the workflow since we have a list of allowed images that can run as
tests I expanded the list accordingly and I moved some images that were
doing easy tasks for already allowed images.

For example we had some debian images that were running `sleep` or
`cat`, such commands can run from a busybox and it was already allowed
so I moved the workflow to use it.

As follow up I will update images to their up to date version.

Signed-off-by: Gianluca Arbezzano <ciao@gianarb.it>