chore: change no-test label usage
We decided to get back to `no-test` label because it is easy to check
for the presence of a single label and we use the content of such label
to explain why we want this example to skip validation. For example we
have set `flaky`, `duration`, `expected-failure` and so on.

Signed-off-by: Gianluca Arbezzano <ciao@gianarb.it>