fix: permission error for executor plugin example doc #13015
The Executor Plugin example documentation does not run correctly as it
is today with the following error:

```
PermissionError: [Errno 13] Permission denied: '/var/run/argo/token'
```

The user set to run the python plugin is incorrect. It should be
`1000` for the particular image we use in the example.

Dejan Golja identified and fixed the issue in its own PR, still open
today:

https://github.com/argoproj/argo-workflows/pull/13019

I tried to rebase it on top of main but the page changed a bit and it is
hard to merge conflicts today.

With this fix the examples works correctly

Co-Authored-By: Dejan Golja <dejan@golja.org>
Signed-off-by: Gianluca Arbezzano <ciao@gianarb.it>